+++ spec/factories/services.rb
@@ -24,7 +24,7 @@ FactoryBot.define do
     sequence(:languages) { |n| "service #{n} lanuage" }
     sequence(:dedicated_for) { |n| ["service #{n} dedicated for"] }
     sequence(:restrictions) { |n| "service #{n} restrictions" }
-    sequence(:phase) { |n| "service #{n} phase" }
+    sequence(:phase) { :alpha }
     sequence(:research_areas) { |n| [create(:research_area)] }
     sequence(:providers) { |n| [create(:provider)] }
     sequence(:categories) { |n| [create(:category)] }